Leverage rapace improvements: Arc<T> blanket impl and Facet-serializable errors
This PR adopts improvements from rapace PRs #61 and #62:

1. **Use Arc<T> blanket impl (Issue #19)**
   - Remove AetherHandle newtype wrapper (~50 lines of boilerplate)
   - Implement Aether trait directly on AetherService
   - Use Arc<AetherService> with server via blanket impl

2. **Leverage Facet-serializable RpcError (Issues #21, #22)**
   - Replace error.to_string() with Arc<RpcError> to preserve context
   - Update vx-rhea and vx-aether error types
   - Full error serialization across RPC boundaries
   - Filed rapace#72 to request native Clone support

3. **Verify RpcSession.run() spawns (Issue #20)**
   - Audited all RpcSession usages - all properly spawn demux loops

Related: rapace#61, rapace#62, rapace#72
Closes: #19, #20, #21, #22
